+++ app/page.tsx
@@ -1,114 +1,136 @@
-import { useCallback, useEffect, useRef, useState } from "react";
+import Link from 'next/link';
+import { getBlogPosts } from '@/db/blog';
 import Image from "next/image";
-import Link from "next/link";
 import Background from "@/components/background";
 
-const paths = [
-    {
-        title: "SWE",
-        link: "/dev",
-        description: "My journey as a Software Developer and Programmer.",
-    },
-    {
-        title: "Photography",
-        link: "/photography",
-        description: "Whether it's traveling or capturing the moment, my photography reveals the beauty in every scene.",
-    },
-]
-
-const store = [
-    {
-        title: "Wallpapers",
-        link: "/store",
-        description: "Download my wallpapers for Phone and Desktop.",
-    },
-]
+const projects = [
+	{
+		repo: "Reacton",
+		link: "https://github.com/jessedoka/reacton",
+		description: "Reaction Tester built with React",
+	},
+	{
+		repo: "InkByter",
+		link: "https://github.com/jessedoka/InkByter",
+		description: "Simple Text Editor made in C",
+	},
+	{
+		repo: "LCT",
+		link: "https://github.com/jessedoka/LCT",
+		description: "A tool for building and managing sentiment analysis lexicons.",
+	},
+	{
+		repo: "Rankhacker",
+		link: "https://github.com/jessedoka/rankhacker",
+		description: "A place where I put all the programming challenges I do",
+	},
+	{
+		repo: "Sudoka",
+		link: "https://github.com/jessedoka/sudoka",
+		description: "A Sudoku game with REST API",
+	},
+	{
+		repo: "Unkbot",
+		link: "https://github.com/jessedoka/unkbot",
+		description: "A Discord bot that just says unk",
+	},
+];
 
 export default function Page() {
+	let allBlogs = getBlogPosts();
 
-    return (
-        <div className="antialiased max-w-2xl md:flex-row mx-4 mt-8 lg:mx-auto d-flex flex-column min-vh-100 mb-auto">
-            <div className="flex-auto min-w-0 mt-6 flex flex-col px-2 md:px-0 mb-auto">
-                <div className="relative w-full h-full">
-                    <div className="absolute inset-0">
-                        <Background width={800} height={600} />
-                    </div>
-                    <div className="absolute bottom-[-14rem] left-4 z-10 ">
-                        <Image src="/icon.svg" alt="icon" width={120} height={120} />
-                    </div>
-                </div>
+	return (
+		<section>
+			<div className="relative w-full h-full">
+				<div className="absolute inset-0">
+					<Background width={800} height={600} />
+				</div>
+				<div className="absolute bottom-[-14rem] left-4 z-10 ">
+					<Image src="/icon.svg" alt="icon" width={120} height={120} />
+				</div>
+			</div>
+			<div className="flex flex-col space-y-4 mt-[15rem]">
+				<div className="space-y-2 mb-4 ">
+					<div>
+						<h1 className="text-2xl text-neutral-800 dark:text-neutral-400 font-medium">
+							Hey there, I&apos;m <span className="text-neutral-800 dark:text-neutral-200 ">Jesse</span> <span className="hover:animate-pulse">👋</span> Welcome to my digital Home 🏡
+						</h1>
+						<h1 className="text-2xl text-neutral-800 dark:text-neutral-400 font-medium">I love to build things and share my experiences.</h1>
+					</div>
+				</div>
+				<h1 className="text-neutral-800 dark:text-neutral-400 text-xl font-medium">
+					My <a href="https://github.com/jessedoka" className='hover:text-orange-400 hover:underline duration-300 transition'>journey</a> has taken me through <Link href="/dev/work" className='hover:text-orange-400 hover:underline duration-300 transition'>internships</Link> and projects where I have developed web applications, enhanced user interactions, and explored the realms of machine learning and predictive analytics.
+				</h1>
+				{/* projects */}
+				<div className='space-y-8'>
+					<h2 className="text-2xl font-medium text-neutral-900 dark:text-neutral-100">
+						Projects
+					</h2>
+					<div className="grid grid-cols-2 gap-4 md:grid-cols-3">
+						{projects.map((project) => (
+							<Link
+								href={project.link}
+								target="_blank"
+								rel="noreferrer"
+								key={project.repo}
+								className="rounded border border-neutral-800 hover:bg-neutral-900 p-4 duration-500 transition-all cursor-pointer"
+							>
+								<div className="flex flex-row space-x-2 items-center">
 
-
-                <div className="mt-[15rem]">
-                    <div className="space-y-2 mb-4 ">
-                        <div>
-                            <h1 className="text-2xl text-neutral-800 dark:text-neutral-400 font-medium">
-                                Hey there, I&apos;m <span className="text-neutral-800 dark:text-neutral-200 ">Jesse</span> <span className="hover:animate-pulse">👋</span> Welcome to my digital Home 🏡
-                            </h1>
-                            <h1 className="text-2xl text-neutral-800 dark:text-neutral-400 font-medium">I love to build things and share my experiences.</h1>
-                        </div>
-
-
-                        <h1 className="text-2xl text-neutral-800 dark:text-neutral-400 font-medium">
-                            Beyond building, I make posters and try to capture moments through my photography that showcase the beauty of the world (I also play bass 🐟).
-                        </h1>
-                    </div>
-
-
-                    <div className="flex flex-col space-y-2">
-                        <h3 className=" text-neutral-400">Paths</h3>
-                        <div className="flex gap-4 justify-between relative">
-                            {paths.map((path) => (
-                                <Link
-                                    href={path.link}
-                                    target="_blank"
-                                    rel="noreferrer"
-                                    key={path.title}
-                                    className="flex-grow flex-shrink-0 rounded border border-neutral-800 hover:bg-neutral-900 p-4 duration-500 transition-all cursor-pointer w-8"
-                                >
-                                    <div className="flex flex-row space-x-2 items-center">
-
-                                        {/* <SiGithub className="w-6 h-6" /> */}
-                                        <p className="text font-medium">
-                                            {path.title}
-                                        </p>
-                                    </div>
-                                    <p className="text-sm text-neutral-700 dark:text-neutral-300">
-                                        {path.description}
-                                    </p>
-                                </Link>
-                            ))}
-                        </div>
-
-                        <h3 className="text-neutral-400 mb-4">Store</h3>
-                        <div className="flex gap-4 justify-between relative">
-                            {store.map((store) => (
-                                <div
-                                    key={store.title}
-                                    className="flex-grow flex-shrink-0 rounded border border-neutral-800 hover:bg-neutral-900 p-4 duration-500 transition-all cursor-pointer w-8"
-                                >
-                                    <div className="flex flex-row space-x-2 items-center">
-                                        <Link
-                                            href={store.link}
-                                            target="_blank"
-                                            rel="noreferrer"
-                                            className="flex items-center space-x-2"
-                                        >
-                                            {/* <SiGithub className="w-6 h-6" /> */}
-                                            <p className="text font-medium">
-                                                {store.title}
-                                            </p>
-                                        </Link>
-                                    </div>
-                                    <p className="text-sm text-neutral-700 dark:text-neutral-300">
-                                        {store.description}
-                                    </p>
-                                </div>
-                            ))}
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    );
+									{/* <SiGithub className="w-6 h-6" /> */}
+									<p className="text font-medium">
+										{project.repo}
+									</p>
+								</div>
+								<p className="text-sm text-neutral-700 dark:text-neutral-300">
+									{project.description}
+								</p>
+							</Link>
+						))}
+					</div>
+				</div>
+				{/* blog list */}
+				<div className="flex flex-col space-y-4 pt-8">
+					<h2 className="text-2xl font-medium text-neutral-900 dark:text-neutral-100">
+						Blog
+					</h2>
+					{allBlogs
+						.sort((a, b) => {
+							if (
+								new Date(a.metadata.publishedAt) > new Date(b.metadata.publishedAt)
+							) {
+								return -1;
+							}
+							return 1;
+						})
+						.slice(0, 5)
+						.map((post) => (
+							<Link
+								key={post.slug}
+								className="flex flex-col space-y-1 mb-4"
+								href={`/blog/${post.slug}`}
+							>
+								<div className="w-full flex flex-col">
+									<div className='flex flex-row space-x-2 justify-between'>
+										<p className="text-neutral-800 dark:text-neutral-300 tracking-tight underline decoration-neutral-400 hover:decoration-neutral-500 hover:">
+											{post.metadata.title}
+										</p>
+										<p className='text-neutral-700 dark:text-neutral-300 text-sm'>
+											{new Date(post.metadata.publishedAt).toLocaleDateString(
+												'en-gb',
+												{
+													year: 'numeric',
+													month: 'short',
+													day: 'numeric',
+												}
+											)}
+										</p>
+									</div>
+								</div>
+							</Link>
+						))}
+				</div>
+			</div>
+		</section >
+	);
 }
