--- app/photography/blog/[slug]/page.tsx
@@ -1,129 +1,0 @@
-
-import type { Metadata } from 'next'
-import { notFound } from 'next/navigation';
-import { CustomMDX } from '@/components/mdx';
-import { getBlogPosts } from '@/db/blog';
-
-export async function generateMetadata({
-	params,
-}:{
-	params: { slug: string };
-}): Promise<Metadata | undefined> {
-	let post = getBlogPosts().find((post) => post.slug === params.slug);
-	if (!post) {
-		return;
-	}
-
-	let {
-		title,
-		publishedAt: publishedTime,
-		summary: description,
-		image,
-	} = post.metadata;
-	let ogImage = image
-		? `https://jessedoka.co${image}`
-		: `https://jessedoka.co/og?title=${title}`;
-
-	return {
-		title,
-		description,
-		openGraph: {
-			title,
-			description,
-			type: 'article',
-			publishedTime,
-			url: `https://jessedoka.co/blog/${post.slug}`,
-			images: [
-				{
-					url: ogImage,
-				},
-			],
-		},
-		twitter: {
-			card: 'summary_large_image',
-			title,
-			description,
-			images: [ogImage],
-		},
-	};
-}
-
-function formatDate(date: string) {
-	let currentDate = new Date();
-	if (!date.includes('T')) {
-		date = `${date}T00:00:00`;
-	}
-	let targetDate = new Date(date);
-
-	let yearsAgo = currentDate.getFullYear() - targetDate.getFullYear();
-	let monthsAgo = currentDate.getMonth() - targetDate.getMonth();
-	let daysAgo = currentDate.getDate() - targetDate.getDate();
-
-	let formattedDate = '';
-
-	if (yearsAgo > 0) {
-		formattedDate = `${yearsAgo}y ago`;
-	} else if (monthsAgo > 0) {
-		formattedDate = `${monthsAgo}mo ago`;
-	} else if (daysAgo > 0) {
-		formattedDate = `${daysAgo}d ago`;
-	} else {
-		formattedDate = 'Today';
-	}
-
-	let fullDate = targetDate.toLocaleString('en-gb', {
-		month: 'long',
-		day: 'numeric',
-		year: 'numeric',
-	});
-
-	return `${fullDate} (${formattedDate})`;
-}
-
-export default function Blog({ params }: { params: { slug: string } }) {
-
-	let post = getBlogPosts().find((post) => post.slug === params.slug);
-
-	if (!post) {
-		notFound();
-		return null;
-	}
-
-	return (
-		<section className='antialiased max-w-2xl md:flex-row mx-4 mt-8 lg:mx-auto d-flex flex-column min-vh-100 mb-auto'>
-			<script
-				type="application/ld+json"
-				suppressHydrationWarning
-				dangerouslySetInnerHTML={{
-					__html: JSON.stringify({
-						'@context': 'https://schema.org',
-						'@type': 'BlogPosting',
-						headline: post.metadata.title,
-						datePublished: post.metadata.publishedAt,
-						dateModified: post.metadata.publishedAt,
-						description: post.metadata.summary,
-						image: post.metadata.image
-							? `https://jessedoka.co${post.metadata.image}`
-							: `https://jessedoka.co/og?title=${post.metadata.title}`,
-						url: `https://jessedoka.co/blog/${post.slug}`,
-						author: {
-							'@type': 'Person',
-							name: 'Jesse Doka',
-						},
-					}),
-				}}
-			/>
-			<h1 className="title font-medium text-2xl tracking-tighter max-w-[650px]">
-				{post.metadata.title}
-			</h1>
-			<div className="flex justify-between items-center mt-2 mb-8 text-sm max-w-[650px]">
-				<p className="text-sm text-neutral-600 dark:text-neutral-400">
-					{formatDate(post.metadata.publishedAt)}
-				</p>
-			</div>
-			<article className="prose prose-quoteless prose-neutral dark:prose-invert">
-				<CustomMDX source={post.content} />
-			</article>
-		</section>
-	);
-}
