--- app/photography/contact/page.tsx
@@ -1,2 +1,0 @@
-export default function ContactPage(){
-}
