--- app/store/page.tsx
@@ -1,8 +1,0 @@
-import React from 'react'
-export default function Store() {
-    return (
-        <div>
-            Store
-        </div>
-    )
-}
